+++ client/src/main/java/com/paraguacraft/pvp/ParaguacraftPvP.java
@@ -36,7 +36,7 @@ public class ParaguacraftPvP {
 
     public static final String MODID = "paraguacraftpvp";
     public static final String NAME = "Paraguacraft PvP Client";
-    public static final String VERSION = "2.1.29";
+    public static final String VERSION = "2.1.31";
 
     @Mod.Instance(MODID)
     public static ParaguacraftPvP instance;