+++ launcher/src-tauri/src/core/loaders/pvp.rs
@@ -33,13 +33,13 @@ const MANIFEST_MIRROR_URLS: &[&str] = &[
     "https://cdn.jsdelivr.net/gh/SantiJ10/Paraguacraft@main/clientes/paraguacraft-pvp/manifest.json",
 ];
 
-const FALLBACK_CLIENT_VERSION: &str = "2.1.30";
-const FALLBACK_RELEASE_TAG: &str = "pvp-client-2.1.30";
+const FALLBACK_CLIENT_VERSION: &str = "2.1.31";
+const FALLBACK_RELEASE_TAG: &str = "pvp-client-2.1.31";
 
 const FALLBACK_MODS: &[(&str, &str)] = &[
     (
-        "ParaguacraftPvP-2.1.30.jar",
-        "2ba566a4f122fbc6809ecf6996891904838034e4",
+        "ParaguacraftPvP-2.1.31.jar",
+        "a5f487acdaec9bb7eda73d8cedf0f357ef321718",
     ),
     (
         "Hytils-Reborn-1.8.9-forge-1.7.5.jar",